Fix points-to SCC collapsing bug
When points-to analysis finds SCCs it marks the wrong node as being
part of a found cycle.  It only wants to mark the node it collapses
to but marked the entry node found rather than the one it collapses
to.  This causes fallout in the patch for PR115236 but generally
weakens the points-to solution by collapsing too many nodes.  Note
that this fix might slow down points-to solving.

	* tree-ssa-structalias.cc (scc_visit): Mark the node we
	collapse to as being in a component.