ada: Fix internal error on case expression used as index of array component
This occurs when the bounds of the array component depend on a discriminant
and the component reference is not nested, that is to say the component is
not (referenced as) a subcomponent of a larger record.

In this case, Analyze_Selected_Component does not build the actual subtype
for the component, but it turns out to be required for constructs generated
during the analysis of the case expression.

The change causes this actual subtype to be built, and also renames a local
variable used to hold the prefix of the selected component.

gcc/ada/

	* sem_ch4.adb (Analyze_Selected_Component): Rename Name into Pref
	and use Sel local variable consistently.
	(Is_Simple_Indexed_Component): New predicate.
	Call Is_Simple_Indexed_Component to determine whether to build an
	actual subtype for the component.