tree-optimization/114435 - pcom left around copies confusing SLP
The following arranges for the pre-SLP vectorization scalar cleanup
to be run when predictive commoning was applied to a loop in the
function.  This is similar to the complete unroll situation and
facilitating SLP vectorization.  Avoiding the SSA copies in predictive
commoning itself isn't easy (and predcom also sometimes unrolls,
asking for scalar cleanup).

	PR tree-optimization/114435
	* tree-predcom.cc (tree_predictive_commoning): Queue
	the next scalar cleanup sub-pipeline to be run when we
	did something.

	* gcc.dg/vect/bb-slp-pr114435.c: New testcase.