Fix some opindex for some options [PR115022]
While looking at the index I noticed that some options had
`-` in the front for the index which is wrong. And then
I noticed there was no index for `mcmodel=` for targets or had
used `-mcmodel` incorrectly.

This fixes both of those and regnerates the urls files see that
`-mcmodel=` option now has an url associated with it.

gcc/ChangeLog:

	PR target/115022
	* doc/invoke.texi (fstrub=disable): Fix opindex.
	(minline-memops-threshold): Fix opindex.
	(mcmodel=): Add opindex and fix them.
	* common.opt.urls: Regenerate.
	* config/aarch64/aarch64.opt.urls: Regenerate.
	* config/bpf/bpf.opt.urls: Regenerate.
	* config/i386/i386.opt.urls: Regenerate.
	* config/loongarch/loongarch.opt.urls: Regenerate.
	* config/nds32/nds32-elf.opt.urls: Regenerate.
	* config/nds32/nds32-linux.opt.urls: Regenerate.
	* config/or1k/or1k.opt.urls: Regenerate.
	* config/riscv/riscv.opt.urls: Regenerate.
	* config/rs6000/aix64.opt.urls: Regenerate.
	* config/rs6000/linux64.opt.urls: Regenerate.
	* config/sparc/sparc.opt.urls: Regenerate.

Signed-off-by: Andrew Pinski <quic_apinski@quicinc.com>