Plugins: Add label-text.h to CPPLIB_H so it will be installed [PR115288]
After r15-874-g9bda2c4c81b668, out of tree plugins won't compile
as the new libcpp header file label-text.h is not installed.

This adds the new header file to CPPLIB_H which is used for
the plugin headers to install.

Committed as obvious after a build and install and make sure
the new header file is installed.

gcc/ChangeLog:

	PR plugins/115288
	* Makefile.in (CPPLIB_H): Add label-text.h.

Signed-off-by: Andrew Pinski <quic_apinski@quicinc.com>