Remove one_if_conv for latest Intel processors.
The tune is added by PR79390 for SciMark2 on Broadwell.
For latest GCC, with and without the -mtune-ctrl=^one_if_conv_insn.
GCC will generate the same binary for SciMark2. And for SPEC2017,
there's no big impact for SKX/CLX/ICX, and small improvements on SPR
and later.

gcc/ChangeLog:

	* config/i386/x86-tune.def (X86_TUNE_ONE_IF_CONV_INSN): Remove
	latest Intel processors.

	Co-authored by: Lingling Kong <lingling.kong@intel.com>